- Rename dist_adam_resamp2 to rpgd - Ensure backwards compatibility
